+++ core-jdk6/src/main/kotlin/uy/klutter/core/jdk/ChildFirstClassloader.kt
@@ -1,12 +1,15 @@
 package uy.klutter.core.jdk
 
+import java.io.InputStream
 import java.net.URL
 import java.net.URLClassLoader
+import java.util.*
 
 // borrowed from KTOR
-// https://github.com/Kotlin/ktor/blob/fd99512bf8e207fef7af5ae6521f871ea9d2fa7b/ktor-core/src/org/jetbrains/ktor/host/OverridingClassLoader.kt
+//   https://github.com/Kotlin/ktor/blob/fd99512bf8e207fef7af5ae6521f871ea9d2fa7b/ktor-core/src/org/jetbrains/ktor/host/OverridingClassLoader.kt
 // probably originally from:
-// https://dzone.com/articles/java-classloader-handling
+//   https://dzone.com/articles/java-classloader-handling
+// then updated to fix resource loading
 
 /**
  * A parent-last classloader that will try the child classloader first and then the parent.
@@ -24,6 +27,20 @@ class ChildFirstClassloader(classpath: List<URL>, parentClassLoader: ClassLoader
             return super.loadClass(name, resolve)
         }
     }
+
+    override fun loadClass(name: String): Class<*> {
+        return loadClass(name, false)
+    }
+
+    override fun findClass(name: String): Class<*> {
+        try {
+            return childClassLoader.findClass(name)
+        } catch (e: ClassNotFoundException) {
+            /* nop */
+        }
+        return super.findClass(name)
+    }
+
 
     /**
      * This class delegates (child then parent) for the findClass method for a URLClassLoader.
@@ -44,4 +61,65 @@ class ChildFirstClassloader(classpath: List<URL>, parentClassLoader: ClassLoader
             }
         }
     }
+
+    override fun findResource(name: String): URL? {
+        return childClassLoader.findResource(name) ?: super.findResource(name)
+    }
+
+    override fun getResource(name: String): URL? {
+        return childClassLoader.getResource(name) ?: super.getResource(name)
+    }
+
+    class UrlEnumeration(val iter: Iterator<URL>): Enumeration<URL> {
+        override fun nextElement(): URL {
+            return iter.next()
+        }
+
+        override fun hasMoreElements(): Boolean {
+            return iter.hasNext()
+        }
+
+    }
+
+    override fun findResources(name: String): Enumeration<URL> {
+        val combined = childClassLoader.findResources(name).asSequence() + super.findResources(name).asSequence()
+        return UrlEnumeration(combined.iterator())
+    }
+
+    override fun getResources(name: String): Enumeration<URL> {
+        val combined = childClassLoader.getResources(name).asSequence() + super.getResources(name).asSequence()
+        return UrlEnumeration(combined.iterator())
+    }
+
+    override fun getResourceAsStream(name: String): InputStream? {
+        return childClassLoader.getResourceAsStream(name) ?: super.getResourceAsStream(name)
+    }
+
+    override fun setClassAssertionStatus(className: String?, enabled: Boolean) {
+        childClassLoader.setClassAssertionStatus(className, enabled)
+        super.setClassAssertionStatus(className, enabled)
+    }
+
+    override fun clearAssertionStatus() {
+        childClassLoader.clearAssertionStatus()
+        super.clearAssertionStatus()
+    }
+
+    override fun setDefaultAssertionStatus(enabled: Boolean) {
+        childClassLoader.setDefaultAssertionStatus(enabled)
+        super.setDefaultAssertionStatus(enabled)
+    }
+
+    override fun setPackageAssertionStatus(packageName: String, enabled: Boolean) {
+        childClassLoader.setPackageAssertionStatus(packageName, enabled)
+        super.setPackageAssertionStatus(packageName, enabled)
+    }
+
+
+    // TODO: should these be overriden?
+    //  override fun getPackages(): Array<out Package>? { }
+    //  override fun getPackage(name: String?): Package? { }
+    //  override fun definePackage(name: String?, specTitle: String?, specVersion: String?, specVendor: String?, implTitle: String?, implVersion: String?, implVendor: String?, sealBase: URL?): Package? { }
+    //  override fun addClass(c: Class<*>?) { }
+    //  override fun getClassLoadingLock(className: String?): Any? { }
 }